RSSI is an int, not a byte

+++ gizmo-fw.ino
@@ -61,7 +61,7 @@ struct BoardState {
   int VBat;
   int WatchdogRemaining;
   int WifiReconnects;
-  byte RSSI;
+  int RSSI;
 
   bool WatchdogOK;
   bool PwrBoard;
